+++ packages/components/src/FormatFile/FormatFile.test.tsx
@@ -1,5 +1,11 @@
 import React from "react";
-import { act, cleanup, fireEvent, render } from "@testing-library/react";
+import {
+  act,
+  cleanup,
+  fireEvent,
+  render,
+  waitFor,
+} from "@testing-library/react";
 import { FormatFile } from ".";
 
 afterEach(cleanup);
@@ -28,9 +34,41 @@ it("renders an image when provided as src", async () => {
     src: () => Promise.resolve(url),
   };
 
-  const { findByRole } = render(<FormatFile file={testFile} />);
+  const { getByAltText } = render(<FormatFile file={testFile} />);
 
-  expect(await findByRole("img")).toBeInTheDocument();
+  expect(getByAltText(testFile.name)).not.toHaveAttribute("src");
+  fireEvent.load(getByAltText(testFile.name));
+
+  await waitFor(() => {
+    expect(getByAltText(testFile.name)).toHaveAttribute("src", url);
+  });
+});
+
+it("renders a skeleton loader when the provided image has not fully loaded", async () => {
+  const url = "not_actually_a_url";
+  const testFile = {
+    key: "234",
+    name: "Onion",
+    type: "image/png",
+    size: 102432,
+    progress: 1,
+    src: () => Promise.resolve(url),
+  };
+
+  const { getByTestId, queryByTestId, getByAltText } = render(
+    <FormatFile file={testFile} />,
+  );
+  expect(getByTestId("internalThumbnailImageLoader")).toBeInTheDocument();
+  expect(getByAltText(testFile.name)).toHaveClass("hidden");
+
+  fireEvent.load(getByAltText(testFile.name));
+
+  await waitFor(() => {
+    expect(
+      queryByTestId("internalThumbnailImageLoader"),
+    ).not.toBeInTheDocument();
+  });
+  expect(getByAltText(testFile.name)).not.toHaveClass("hidden");
 });
 
 it("should call the delete handler", async () => {
@@ -80,7 +118,7 @@ describe("when the format file is a thumbnail", () => {
 
       const thumbnailImage = await findByRole("img");
 
-      expect(thumbnailImage.parentElement.className).toContain("base");
+      expect(thumbnailImage.parentElement?.className).toContain("base");
     });
   });
 
@@ -102,7 +140,7 @@ describe("when the format file is a thumbnail", () => {
 
       const thumbnailImage = await findByRole("img");
 
-      expect(thumbnailImage.parentElement.className).toContain("large");
+      expect(thumbnailImage.parentElement?.className).toContain("large");
     });
   });
 });